Fix Build function, Resolves #21 (#23)
* Fix build function (#21)

The build function was using a bad process creation method. We changed to a better method and started printing the build outputs to files instead of to stdout.

Co-authored-by: Agriad <agriad1@yahoo.com>
Co-authored-by: davek10 <94davek@gmail.com>
Co-authored-by: tjr16 <jerryaccount@126.com>

* Refactor building and add test (#21)

Removed the parse build since our new take on the
building does not require it. Also updated the test to address that.

Co-authored-by: Agriad <agriad1@yahoo.com>
Co-authored-by: davek10 <94davek@gmail.com>
Co-authored-by: tjr16 <jerryaccount@126.com>

* Add project for testing purposes (#21)

Co-authored-by: Agriad <agriad1@yahoo.com>
Co-authored-by: davek10 <94davek@gmail.com>
Co-authored-by: tjr16 <jerryaccount@126.com>

Co-authored-by: Markus Wesslén <markus.wesslen@gmail.com>
Co-authored-by: Agriad <agriad1@yahoo.com>
Co-authored-by: tjr16 <jerryaccount@126.com>